+++ uiscanner/src/main/java/no/nordicsemi/android/common/ui/scanner/repository/DevicesRepository.kt
@@ -29,7 +29,7 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.android.common.ui.scanner.scanner.repository
+package no.nordicsemi.android.common.ui.scanner.repository
 
 import dagger.hilt.android.scopes.ViewModelScoped
 import kotlinx.coroutines.channels.awaitClose
